+++ wemake_python_styleguide/visitors/presets/tokens.py
@@ -7,6 +7,7 @@ from wemake_python_styleguide.visitors.tokenize.primitives import (
     WrongPrimitivesVisitor,
 )
 
+#: Used to store all token related visitors to be later passed to checker:
 TOKENS_PRESET = (
     WrongCommentVisitor,
     WrongPrimitivesVisitor,